shopping cart and order module DTO updates

+++ OrderProcessingModule/src/main/java/com/cs490/shoppingCart/OrderProcessingModule/repository/OrderRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order,Integer> {
-    List<Order> findAllByUserIdEquals(int userId);
+public interface OrderRepository extends JpaRepository<Order,Long> {
+    List<Order> findAllByUserIdEquals(Long userId);
 }
