+++ ShoppingCartModule/src/main/java/ecommerce/shoppingcartservice/repository/CartLineRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartLineRepository extends JpaRepository<CartLine,Integer> {
+public interface CartLineRepository extends JpaRepository<CartLine,Long> {
 
 
 }
